Window: ask for confirmation before closing with unsaved changes
This adds another property to the perspectives, which signals whether a
perspective can safely shutdown. On a delete event, the Window checks
all its perspectives' properties and if one perspective signals that it
cannot safely shutdown, a dialog is presented that allows the user to
either shutdown ignoring any changes or to abort and commit or make other changes.

Fixes #102