Introduce RatbagdMacro to abstract macros
This change cleans up the shady KeyStroke class that never felt quite
right. Its behavior is split over RatbagdMacro, which is a new class
abstracting a macro as a list of (type, keycode) tuples with a
human-readable representation and methods to manipulate said list, and
the existing ButtonDialog class which is now again fully responsible for
any and all input event handling. RatbagdMacro has no notion whatsoever
of Gdk; all of this is now only found in Piper's GUI code where it
belongs.