Add the ErrorPerspective before all others
If ratbag is None, we crash during add_perspective because the other
perspectives rely on it being valid. Instead of adding a bunch of "is None" to
the add_perspective call, just add the ErrorPerspective first, check for
ratbag and exit if it doesn't exist. Otherwise, add the other two
perspectives and continue on our merry way.