RatbagdProfile: emit notify::is-active from DBus' PropertyChanged signal
This ensures that it is fired in all instances where the property is
updated, and not just on our own `set_active` method.

We also cache the property so that we can emit the signals only for the
profile that *was* the active one, and the one that *became* the active
one, to circumvent
https://github.com/libratbag/libratbag/blob/bb23f68d53a5da85c195def66ec103f086b9254b/ratbagd/ratbagd-profile.c#L271

Fixes #167.