Do not generate icon-cache when DESTDIR is set
This ensures that the icon cache will be generated when installed locally,
but will also make sure that we don't generate one when DESTDIR is set.
This helps with packaging, where packages should never provide the duplicate
"icon-theme.cache" file inside the package.

This is based around the meson build system implemented by @TingPing in
the gnome-twitch project.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>