data: add the appdata.xml file
Can be tested by installing the .xml into /usr/share/metainfo/, then running
gnome-software --prefer-local (after killing any running g-s instance). This
may take a while to refresh and if you have a repo that contains piper it may
still be preferred for some reason.

Piper is under "Utilities", if the search doesn't find it immediately, wait
for a while? Or start searching for "org.freedesktop.Piper".

Technically appdata requires 16:9 screenshots but it seems to work with the
ones we have online for the wiki already (they're a little bit too high for
16:9).

Fixes #236

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>