Hide the "add a resolution" button
This hasn't been implemented in over a year, let's hide it from sight. Maybe
that spurns someone on to actually implement it.

It's an improvement to the GUI either way because the message printed to stdout
wouldn't be seen by most users starting the Piper through e.g. gnome shell. So
it just looks like the button is broken.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>