Hook up SVG loading from gresources
libratbag will stop providing the SVGs, so let's load them from ourselves. To
do so we rely on the new DBus property "UsbId" and a lookup table
(svg-lookup.ini).

That file was generated with:

    for file in `grep -l "Svg=" *.device`; do
        grep -h -E "Name|Match|Svg" $file  | sed -e "s/Name=\(.*\)/\n[\1]/g" >> outfile.ini;
    done

The SVGs are added to our gresource bundle and loaded from there. A helper
function takes care of this.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>