Check for incompatible ratbagd API versions
Instead of just crashing or not working when ratbagd isn't the version we
expect, let's check for the version. This shows version -1 for those ratbagd
versions where we don't have the field but meh.

Since this one is thrown in the constructor, let's pass a function to create
ratbagd up to the Window from the Application. Then we can call that and
immediately bind the exceptions into our error perspectives.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>